move percentage to common utils crate
