+++ crates/api_models/tests/percentage.rs
@@ -1,6 +1,5 @@
 #![allow(clippy::panic_in_result_fn)]
-use api_models::types::Percentage;
-use common_utils::errors::ApiModelsError;
+use common_utils::{errors::ApiModelsError, types::Percentage};
 const PRECISION_2: u8 = 2;
 const PRECISION_0: u8 = 0;
